Move common logic to base synchronizer
